requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# Optional: For better JSON handling
pydantic>=2.5.0

# Optional: For evaluation and plotting
pandas>=2.1.0
matplotlib>=3.8.0
seaborn>=0.13.0
